% PKPDx/man/pk1dt.Rd
% Part of the PKPDx package for R 
% Distributed under GPL or later: see PKPDx/LICENCE
\name{pk1dt}
\alias{pk1dt}
\docType{data}
\title{pk1dt Data (10 x 5)}% Data Title
\description{Data pk1dt
  ...
}
\usage{data(pk1dt)}
\format{
The \code{pk1dt} data frame has 10 rows and 5 columns
\describe{
   \item{x}{
     a numeric vector with values from 10 to 150
   }
   \item{y1}{
     a numeric vector with values from 200 to 920
   }
   \item{y2}{
     a numeric vector with values from 42 to 850
   }
   \item{y3}{
     a numeric vector with values from 60 to 800
   }
   \item{y4}{
     a numeric vector with values from 120 to 465
   }
}}
\details{
 ...
}
\source{
  ???? Doe, A. B. and Doe, J. M. (xxxx), \emph{Title},
  Springer, New York.  (Appendix A.17)
}
\examples{
summary(pk1dt)
}
\keyword{datasets}
