% PKPDx/man/pk2dt.Rd
% Part of the PKPDx package for R 
% Distributed under GPL or later: see PKPDx/LICENCE
\name{pk2dt}
\alias{pk2dt}
\docType{data}
\title{pk2dt Data (13 x 5)}% Data Title
\description{Data pk2dt
  ...
}
\usage{data(pk2dt)}
\format{
The \code{pk2dt} data frame has 13 rows and 5 columns
\describe{
   \item{time}{
     a numeric vector with values from 10 to 360
   }
   \item{conc}{
     a numeric vector with values from 0 to 2
   }
   \item{civ}{
     a numeric vector with values from 0.2 to 4.7
   }
   \item{inpt}{
     a numeric vector with values from -0.00094 to 0.01932
   }
   \item{cinpt}{
     a numeric vector with values from 0 to 0.646
   }
}}
\details{
 ...
}
\source{
  ???? Doe, A. B. and Doe, J. M. (xxxx), \emph{Title},
  Springer, New York.  (Appendix A.17)
}
\examples{
summary(pk2dt)
}
\keyword{datasets}
